nuevos casos de prueba escritos de la catedra, corregido tipo linteger representado como long en parser, 3er entrega todos los casos testeados

+++ src/ASFinEnteroLargo.java
@@ -6,7 +6,7 @@ public class ASFinEnteroLargo implements AccionSemantica {
         al.setTokenActual(al.getIDforPR("CTE"));
         al.addListaDeTokens(String.format("CTE linteger (linea %1$d)", al.getLinea()));
         //evaluo si es mayor de lo permitido
-        Double valor = Double.valueOf(al.getBuffer().substring(0, al.getBuffer().length() - 1));
+        Long valor = Long.valueOf(al.getBuffer().substring(0, al.getBuffer().length() - 1));
         if (valor > AnalizadorLexico.MAX_LONG) {
             al.addListaDeErroresLexicos(String.format("Warning cte entera fuera de rango at linea: %1$d", al.getLinea()));
             valor = AnalizadorLexico.MAX_LONG;
